Fix labeling, combine into file
